refactor(context_compressor): extract tool-result summarizers (LB5a)
Window 459-634 (def-aligned; pin 1114-1289 per W2 redraw) into
context_compressor_summarizers.py. 3 members re-exported is-identical;
logger dep via lazy round-trip import (auth s2 seam pattern, zero
behavior change); skill-prune deps one-way from skill_prune module.
47/47 targeted tests green.

Part of #78645
Part of #78647
